docs: :fire: Started migration to TS
+ started some doc

Many issues with `firestoreAction` & types imports

+++ store/auth/actions.ts
@@ -1,5 +1,14 @@
-export default {
-  async onAuth({ commit, dispatch }, { authUser }) {
+import type { ActionTree } from 'vuex'
+import type firebase from 'firebase'
+import type { RootState } from '..'
+import type { AuthState } from './state'
+import { Account, InputUser, User } from '~/types'
+
+const actions: ActionTree<AuthState, RootState> = {
+  async onAuth(
+    { commit, dispatch },
+    { authUser }: { authUser: firebase.User | null }
+  ) {
     if (authUser) {
       commit('SET_AUTH_USER', authUser)
       await dispatch('account/bindAccounts', authUser, { root: true })
@@ -7,16 +16,19 @@ export default {
       await dispatch('bindSettings', {}, { root: true })
     } else {
       commit('RESET_STATE')
-      await dispatch('account/unbindAccounts', null, { root: true })
-      await dispatch('agenda/unbindAgenda', null, { root: true })
+      await dispatch('account/unbindAccounts', {}, { root: true })
+      await dispatch('agenda/unbindAgenda', {}, { root: true })
       await dispatch('unbindSettings', {}, { root: true })
     }
   },
-  async createUser({ dispatch }, { email, password, balance }) {
+  async createUser({ dispatch }, { email, password, balance }: InputUser) {
     const { user } = await this.$fire.auth.createUserWithEmailAndPassword(
       email,
       password
     )
+    if (!user) {
+      throw new Error('Une erreur est survenue lors de la création du compte')
+    }
 
     const date = new Date()
 
@@ -30,19 +42,30 @@ export default {
       'account/createAccount',
       {
         name: 'Courant',
-        balance,
-      },
+        balance: parseFloat(balance),
+      } as Account,
       { root: true }
     )
   },
   async deleteUser({ getters, dispatch }) {
-    const { uid } = getters.getUser
+    if (!this.$fire.auth.currentUser) {
+      throw new Error('Une erreur est survenue, essayez de vous reconnecter')
+    }
+    const uid = (getters.getUser as User | null)?.uid
+    if (!uid) {
+      throw new Error('Vous devez être connecté pour effectuer cette action')
+    }
+
     await dispatch('onAuth', { authUser: null })
     const ref = this.$fire.firestore.collection('users').doc(uid)
     await ref.delete()
     await this.$fire.auth.currentUser.delete()
   },
-  async updateUser(ctx, { email, password }) {
+  async updateUser(_, { email, password }) {
+    if (!this.$fire.auth.currentUser) {
+      throw new Error('Une erreur est survenue, essayez de vous reconnecter')
+    }
+
     if (email) {
       await this.$fire.auth.currentUser.updateEmail(email)
       // await this.$fire.auth.currentUser.verifyBeforeUpdateEmail(email)
@@ -62,3 +85,5 @@ export default {
     return this.$fire.auth.signOut()
   },
 }
+
+export default actions
